Ported to RHEL5

+++ apps/vdf2raw/vdf2raw.cpp
@@ -122,7 +122,7 @@ void	process_volume(
 		wbreader2D->CloseVariable();
 
 		if (! opt.quiet) {
-			fprintf(stdout, "Wrote %dx%d plane\n", dim[0],dim[1]);
+			fprintf(stdout, "Wrote %dx%d plane\n", (int) dim[0],(int) dim[1]);
 		}
 		delete wbreader2D;
 		return;
@@ -157,7 +157,7 @@ void	process_volume(
 		wbreader3D->CloseVariable();
 
 		if (! opt.quiet) {
-			fprintf(stdout, "Wrote %dx%dx%d volume\n", dim[0],dim[1],dim[2]);
+			fprintf(stdout, "Wrote %dx%dx%d volume\n", (int) dim[0],(int) dim[1],(int) dim[2]);
 		}
 	}
 	else {
@@ -204,7 +204,7 @@ void	process_volume(
 		if (! opt.quiet) {
 			fprintf(
 				stdout, "Wrote %dx%dx%d volume\n", 
-				bdim[0]*bs[0],bdim[1]*bs[0],bdim[2]*bs[0]
+				(int) (bdim[0]*bs[0]),(int) (bdim[1]*bs[0]),(int) (bdim[2]*bs[0])
 			);
 		}
 	}
@@ -286,8 +286,8 @@ void	process_region(
 		TIMER_STOP(t0, *write_timer);
 
 		if (! opt.quiet) {
-			if(is3D) fprintf(stdout, "Wrote %dx%dx%d volume\n", dim[0],dim[1],dim[2]);
-			else fprintf(stdout, "Wrote %dx%d plane\n", dim[0],dim[1]);
+			if(is3D) fprintf(stdout, "Wrote %dx%dx%d volume\n", (int) dim[0],(int) dim[1],(int) dim[2]);
+			else fprintf(stdout, "Wrote %dx%d plane\n", (int) dim[0],(int) dim[1]);
 		}
 	}
 	else {
@@ -338,11 +338,11 @@ void	process_region(
 		if (! opt.quiet) {
 			if (is3D) fprintf(
 				stdout, "Wrote %dx%dx%d volume\n", 
-				bdim[0]*bs[0],bdim[1]*bs[0],bdim[2]*bs[0]
+				(int) (bdim[0]*bs[0]),(int) (bdim[1]*bs[0]),(int) (bdim[2]*bs[0])
 			);
 			else fprintf(
 				stdout, "Wrote %dx%d plane\n", 
-				bdim[0]*bs[0],bdim[1]*bs[0]
+				(int) (bdim[0]*bs[0]),(int) (bdim[1]*bs[0])
 			);
 		}
 	}
